+++ src/Casters/ArrayCaster.php
@@ -5,6 +5,7 @@ namespace Spatie\DataTransferObject\Casters;
 use ArrayAccess;
 use LogicException;
 use Spatie\DataTransferObject\Caster;
+use Traversable;
 
 class ArrayCaster implements Caster
 {
@@ -14,7 +15,7 @@ class ArrayCaster implements Caster
     ) {
     }
 
-    public function cast(mixed $value): array | ArrayAccess
+    public function cast(mixed $value): array|ArrayAccess
     {
         if ($this->type == 'array') {
             return $this->castArray($value);
@@ -27,23 +28,21 @@ class ArrayCaster implements Caster
         throw new LogicException("Caster [ArrayCaster] may only be used to cast arrays or objects that implement ArrayAccess.");
     }
 
-    private function castArray(array $value): array
+    private function castArray(mixed $value): array
     {
-        return array_map([$this, 'makeItem'], $value);
+        return $this->mapInto(destination: [], items: $value);
     }
 
-    private function castArrayAccess(array $value): ArrayAccess
+    private function castArrayAccess(mixed $value): ArrayAccess
     {
-        $arrayAccess = new $this->type();
-
-        foreach ($this->castArray($value) as $item) {
-            $arrayAccess[] = $item;
+        if (! is_subclass_of($this->type, Traversable::class)) {
+            throw new LogicException("Caster [ArrayCaster] may only be used to cast ArrayAccess objects that are traversable.");
         }
 
-        return $arrayAccess;
+        return $this->mapInto(destination: new $this->type(), items: $value);
     }
 
-    private function makeItem(mixed $data): mixed
+    private function castItem(mixed $data)
     {
         if ($data instanceof $this->itemType) {
             return $data;
@@ -54,7 +53,16 @@ class ArrayCaster implements Caster
         }
 
         throw new LogicException(
-            'Caster [ArrayCaster] requires each item to be either an array or an instance of the specified class.'
+            "Caster [ArrayCaster] each item must be an array or an instance of the specified item type [{$this->itemType}]."
         );
+    }
+
+    private function mapInto(array|Traversable $destination, mixed $items): array|ArrayAccess
+    {
+        foreach ($items as $key => $item) {
+            $destination[$key] = $this->castItem($item);
+        }
+
+        return $destination;
     }
 }